feat: added blogs controller error handling

+++ Talkish.Dal/Repositories/BlogRepository.cs
@@ -26,6 +26,12 @@ namespace Talkish.Dal.Repositories
         public async Task<Blog> DeleteBlogByIdAsync(int Id)
         {
             Blog blogToRemove = await _ctx.Blogs.FirstOrDefaultAsync((b) => b.BlogId == Id);
+
+            if (blogToRemove == null)
+            {
+                return null;
+            }
+
             _ctx.Blogs.Remove(blogToRemove);
             await _ctx.SaveChangesAsync();
             return blogToRemove;
@@ -56,6 +62,12 @@ namespace Talkish.Dal.Repositories
             Blog blog = await _ctx.Blogs
                 .Include((blog) => blog.Topics)
                 .FirstOrDefaultAsync((blog) => blog.BlogId == Id);
+
+            if (blog == null)
+            {
+                return null;
+            }
+
             return blog.Topics;
         }
 
@@ -64,6 +76,12 @@ namespace Talkish.Dal.Repositories
             Blog blog = await _ctx.Blogs.FirstOrDefaultAsync((blog) => blog.BlogId == BlogId);
             blog.Title = BlogData.Title;
             blog.Content = BlogData.Content;
+
+            if (blog == null)
+            {
+                return null;
+            }
+
             _ctx.Blogs.Update(blog);
             await _ctx.SaveChangesAsync();
             return blog;
@@ -78,6 +96,11 @@ namespace Talkish.Dal.Repositories
             Topic topic = await _ctx.Topics
                 .Where((topic) => topic.TopicId == TopicId)
                 .FirstOrDefaultAsync();
+
+            if (blog == null || topic == null)
+            {
+                return null;
+            }
 
             blog.Topics.Add(topic);
             await _ctx.SaveChangesAsync();
